Making progress on generalized GUI handling.
git-svn-id: https://s.snth.net/svn/rlk/thumb@1053 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/matrix.hpp
@@ -129,7 +129,7 @@ void get_plane(double *, const double *, const double *, const double *);
 
 //-----------------------------------------------------------------------------
 
-double nearestint(double);  // TODO: This doesn't belong here.
+int nearestint(double);  // TODO: This doesn't belong here.
 
 void   midpoint(double *, const double *, const double *);
 double distance(          const double *, const double *);
